feat: decouple Transaction and SignedTransaction classes from serialization
