+++ packages/providers/src/json-rpc-provider.ts
@@ -29,6 +29,7 @@ import {
     TypedError,
 } from '@near-js/types';
 import {
+    encodeTransaction,
     SignedTransaction,
 } from '@near-js/transactions';
 import { baseEncode } from 'borsh';
@@ -81,7 +82,7 @@ export class JsonRpcProvider extends Provider {
      * @param signedTransaction The signed transaction being sent
      */
     async sendTransaction(signedTransaction: SignedTransaction): Promise<FinalExecutionOutcome> {
-        const bytes = signedTransaction.encode();
+        const bytes = encodeTransaction(signedTransaction);
         return this.sendJsonRpc('broadcast_tx_commit', [Buffer.from(bytes).toString('base64')]);
     }
 
@@ -92,7 +93,7 @@ export class JsonRpcProvider extends Provider {
      * @returns {Promise<FinalExecutionOutcome>}
      */
     async sendTransactionAsync(signedTransaction: SignedTransaction): Promise<FinalExecutionOutcome> {
-        const bytes = signedTransaction.encode();
+        const bytes = encodeTransaction(signedTransaction);
         return this.sendJsonRpc('broadcast_tx_async', [Buffer.from(bytes).toString('base64')]);
     }
