+++ packages/transactions/src/index.ts
@@ -1,5 +1,6 @@
 export * from './action_creators';
 export * from './actions';
 export * from './create_transaction';
-export * from './schema';
+export * from './serialization';
 export * from './sign';
+export * from './transaction';
